+++ packages/stargate/src/search.ts
@@ -1,11 +1,3 @@
-export interface SearchByHeightQuery {
-  readonly height: number;
-}
-
-export interface SearchBySentFromOrToQuery {
-  readonly sentFromOrTo: string;
-}
-
 /**
  * This query type allows you to pass arbitrary key/value pairs to the backend.
  */